Add JWTRevocationTestCase
